requirements.txt
# 제안요청서 요약 - JupyterLab용
openai>=1.0.0
python-docx>=1.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0

# RAG & 벡터화
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-openai>=0.0.5
pypdf>=3.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
